KIP-297: MINOR: move FileConfigProvider to provider subpackage
