+++ clients/src/main/java/org/apache/kafka/common/config/provider/FileConfigProvider.java
@@ -14,7 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.config;
+package org.apache.kafka.common.config.provider;
+
+import org.apache.kafka.common.config.ConfigData;
+import org.apache.kafka.common.config.ConfigException;
+import org.apache.kafka.common.config.ConfigProvider;
 
 import java.io.BufferedReader;
 import java.io.FileInputStream;
